CX-411: Re-added signed fix to size-1

+++ source/plugins/org.custusx.registration.method.landmark/cxLandmarkRegistrationWidget.cpp
@@ -210,8 +210,8 @@ QString LandmarkRegistrationWidget::getNextLandmark()
 {
 	std::vector<Landmark> lm = this->getAllLandmarks();
 
-	unsigned long size = lm.size();
-	for (unsigned i=0; i < size-1; ++i)
+	int size = int(lm.size());
+	for (unsigned i=0; int(i) < size-1; ++i)
 	{
 		if (lm[i].getUid()==mActiveLandmark)
 		{
